refactor(tnt): share sender id extraction between transient and targeted messages
